refactor: Extract schema utility functions to separate module
