+++ frontend/src/components/nodes/nodeSidebar/NodeSidebar.tsx
@@ -43,6 +43,7 @@ import NodeOutput from '../NodeOutputDisplay'
 import OutputSchemaEditor from './OutputSchemaEditor'
 import SchemaEditor from './SchemaEditor'
 
+import { extractSchemaFromJsonSchema, generateJsonSchemaFromSchema } from '@/utils/schemaUtils'
 import { convertToPythonVariableName } from '@/utils/variableNameUtils'
 import Ajv from 'ajv'
 
@@ -78,75 +79,7 @@ const nodesComparator = (prevNodes: FlowWorkflowNode[], nextNodes: FlowWorkflowN
     return prevNodes.every((node, index) => nodeComparator(node, nextNodes[index]))
 }
 
-// Update the extractSchemaFromJsonSchema function to return an error instead of calling setError
-const extractSchemaFromJsonSchema = (
-    jsonSchema: string
-): { schema: Record<string, any> | null; error: string | null } => {
-    if (!jsonSchema || !jsonSchema.trim()) {
-        return { schema: null, error: null }
-    }
-    try {
-        // Try to parse the schema
-        let parsed: Record<string, any>
-        try {
-            parsed = JSON.parse(jsonSchema.trim())
-        } catch (e: any) {
-            // If the schema has escaped characters, clean it up first
-            let cleaned = jsonSchema
-                .replace(/\"/g, '"') // Replace escaped quotes
-                .replace(/\\\[/g, '[') // Replace escaped brackets
-                .replace(/\\\]/g, ']')
-                .replace(/\\n/g, '') // Remove newlines
-                .replace(/\\t/g, '') // Remove tabs
-                .replace(/\\/g, '') // Remove remaining backslashes
-                .trim()
-            try {
-                parsed = JSON.parse(cleaned)
-            } catch (e: any) {
-                // Extract line and column info from the error message if available
-                const match = e.message.match(/at position (\d+)(?:\s*\(line (\d+) column (\d+)\))?/)
-                const errorMsg = match
-                    ? `Invalid JSON: ${e.message.split('at position')[0].trim()} at line ${match[2] || '?'}, column ${match[3] || '?'}`
-                    : `Invalid JSON: ${e.message}`
-                return { schema: null, error: errorMsg }
-            }
-        }
-
-        // If the parsed schema has a properties field (i.e. full JSON Schema format),
-        // return the nested properties so that nested objects are preserved.
-        if (parsed.properties) {
-            return { schema: parsed.properties, error: null }
-        }
-        return { schema: parsed, error: null }
-    } catch (error: any) {
-        return { schema: null, error: error.message || 'Invalid JSON Schema' }
-    }
-}
-
 // Add this helper function near the top, after extractSchemaFromJsonSchema
-const generateJsonSchemaFromSchema = (schema: Record<string, string>): string | null => {
-    if (!schema || Object.keys(schema).length === 0) return null
-
-    try {
-        const jsonSchema = {
-            type: 'object',
-            required: Object.keys(schema),
-            properties: {} as Record<string, { type: string }>,
-        }
-
-        for (const [key, type] of Object.entries(schema)) {
-            if (!key || !type) return null
-            jsonSchema.properties[key] = { type }
-        }
-
-        return JSON.stringify(jsonSchema, null, 2)
-    } catch (error) {
-        console.error('Error generating JSON schema:', error)
-        return null
-    }
-}
-
-// Add this function after the existing helper functions and before the NodeSidebar component
 const getModelConstraints = (nodeSchema: FlowWorkflowNodeType | null, modelId: string): ModelConstraints | null => {
     if (!nodeSchema || !nodeSchema.model_constraints || !nodeSchema.model_constraints[modelId]) {
         return null
@@ -700,8 +633,6 @@ const NodeSidebar: React.FC<NodeSidebarProps> = ({ nodeID }) => {
                     setCurrentNodeConfig={setCurrentNodeConfig}
                     dispatch={dispatch}
                     updateNodeConfigOnly={updateNodeConfigOnly}
-                    generateJsonSchemaFromSchema={generateJsonSchemaFromSchema}
-                    extractSchemaFromJsonSchema={extractSchemaFromJsonSchema}
                     isLast={isLast}
                 />
             )
